Create Field Sheets

create_field_sheet() appends a worksheet to an existing excel workbook containing a formatted field sheet to be used by crew collecting genetic samples.

create_field_sheet(
  wb,
  field_sheet_sample_plan,
  sample_event_number,
  first_sample_date,
  sample_location,
  sample_location_code
)

Arguments

wb
A Workbook object from createWorkbook

field_sheet_sample_plan
A dataframe containing the content of the field sheet, use the field_sheet_sample_plan output from get_field_sheet_event_plan. Note get_field_sheet_event_plan also returns the other arguments required by this function.

The following columns must contain data:

Bin The sampling bin identifier A-E

Bin FL Range (mm) The min to max fork length range of the sampling bin

Sample # A number denoting the order the sample was taken, 1 to the total number of planned samples with in the sampling bin

Sample ID {sample_location_code}{YY}_{sample_event_number}_{sample_bin}_{sample_number} (e.g., "BTC22_3_A_1")

The remaining columns ("Date", "Time", "FL (mm)", "Field Run ID", "Fin Clip (Y/N)", "Comments") will be left empty and filled out by hand in the field

sample_event_number
The non-unique sampling event number, enumerated from 1 at the start of each monitoring season for each sampling location

first_sample_date
A date object YYYY-MM-DD representing the first day of sampling of a 2 day sampling event

sample_location
The sampling location name (e.g., "Battle Creek")

sample_location_code
The sampling location short code (e.g., "BTC")

Value

A Workbook object from createWorkbook with the new worksheet

See also

Examples

cfg <- config::get()

con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
wb <- openxlsx::createWorkbook()
# Each sample_event will be a tab in a workbook
plan <- get_field_sheet_event_plan(con, sample_event_id = 1)
wb <- create_field_sheet(wb = wb,
                         field_sheet_sample_plan = plan$field_sheet_sample_plan,
                         sample_event_number = plan$sample_event_number,
                         first_sample_date = plan$first_sample_date,
                         sample_location = plan$location_name,
                         sample_location_code = plan$location_code)

#> Error in if (tolower(sheetName) %in% tolower(wb$sheet_names)) {    stop(paste0("A worksheet by the name '", sheetName, "' already exists! Sheet names must be unique case-insensitive."))}: argument is of length zero

plan <- get_field_sheet_event_plan(con, sample_event_id = 2)
wb <- create_field_sheet(wb = wb,
                         field_sheet_sample_plan = plan$field_sheet_sample_plan,
                         sample_event_number = plan$sample_event_number,
                         first_sample_date = plan$first_sample_date,
                         sample_location = plan$location_name,
                         sample_location_code = plan$location_code)

#> Error in if (tolower(sheetName) %in% tolower(wb$sheet_names)) {    stop(paste0("A worksheet by the name '", sheetName, "' already exists! Sheet names must be unique case-insensitive."))}: argument is of length zero

openxlsx::saveWorkbook(wb, "test.xlsx", overwrite = TRUE)

#> Warning: Workbook does not contain any worksheets. A worksheet will be added.