Create Multiple Field Sheets

create_multiple_field_sheets() creates an excel workbook and appends multiple formatted field worksheets for sample event IDs in a given sample plan.

Usage

create_multiple_field_sheets(added_sample_plan, field_sheet_filepath)

Arguments

added_sample_plan

The object created by running add_sample_plan(). This is a named list containing elements "number_of_samples_added" and "sample_ids_created". "sample_ids_created" is a table and must contain a column "sample_event_id".

field_sheet_filepath

The filepath and name desired for the workbook containing field sheets.

Value

A Workbook object from createWorkbook with a worksheet for each sampling event in the sample plan.

Details

create_multiple_field_sheets() combines get_field_sheet_sample_plan() and create_field_sheet() into one function. Field sheets are created for all unique sample event IDs in a given sample plan. get_field_sheet_sample_plan() and create_field_sheet() can still be run independently to create one field sheet at a time.

See also

Examples