Delete Sample Location

delete_sample_location() deletes an existing sample location in the sample location lookup table

delete_sample_location(con, sample_location_id)

Arguments

con
A DBI connection object obtained from DBI::dbConnect()

sample_location_id
A numeric ID for the targeted sample location get_sample_locations

Details

Note: If an sample location has been associated with a permit record, then the database restricts deleting this sample location. You must first update those records with a new sample location before reattempting to delete the sample location. Consider using the update_sample_location_status function if you are wanting to retire an sample location while retaining its value for historic records.

See also

Examples

# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)

all_sample_locations <- get_sample_locations(con)
View(all_sample_locations) # to view the ID of the sample location needing deletion

delete_sample_location(con, 4)

#> Error: Failed to prepare query: ERROR:  operator does not exist: character varying = integer
#> LINE 1: DELETE FROM sample location WHERE id = 4;
#>                                              ^
#> HINT:  No operator matches the given name and argument types. You might need to add explicit type casts.