Get Sample Event Information for Creating a Field Sheet

get_field_sheet_event_plan() retrieves the sampling event information from the database that is needed to prepare field sheets.

Usage

get_field_sheet_event_plan(con, sample_event_id_arg)

Arguments

con

A DBI connection object obtained from DBI::dbConnect()

sample_event_id

The numeric unique identifier of the targeted sampling event for a location Use get_sample_event to query and retrieve the IDs from the database.

Value

A list object containing the required arguments for create_field_sheet

field_sheet_sample_plan A dataframe containing the content of the field sheet

sample_event_number The non-unique sample event ID

first_sample_date A date object YYYY-MM-DD representing the first day of sampling of a 2 day sampling event

location_name The sampling location name (e.g., "Battle Creek")

location_code The sampling location short code (e.g., "BTC")

See also

Examples

cfg <- config::get()

con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
#> Error: FATAL:  The access token has expired. Please acquire a new token and retry.
#> FATAL:  The access token has expired. Please acquire a new token and retry.
plan <- get_field_sheet_event_plan(con, sample_event_id = 1)
#> Error in eval(expr, envir, enclos): object 'con' not found