Get Samples by Season

View sample by season with status and run (if assigned)

Usage

get_samples_by_season(
  con,
  season,
  dataset = c("raw", "clean", "unprocessed"),
  heterozygote_filter = c(FALSE, TRUE),
  failed_filter = c(FALSE, TRUE)
)

Arguments

con

connection to the database

season

year in format YYYY. You can pass in a min and max season as c(YYYY, YYYY). A season consists of all sampling events from the given year up to September 30th and from the previous year after October 1st.

dataset

either "raw", "clean", or "unprocessed".

heterozygote_filter

defaults to FALSE. If TRUE, only heterozygotes are returned.

failed_filter

defaults to FALSE. If TRUE, only "failed" assays (negative for both OTS28 early and late, or negative for both OTS16 spring and winter)

Details

the parameter dataset can be used to determine what information is included in the tibble. The raw, clean, and unprocessed datasets contain the following variables:

stream_name

datetime_collected

sample_event_number

sample_id

sherlock_run_assignment

field_run_assignment

fork_length_mm

fin_clip

status

updated_at

If you select dataset = "raw", the tibble will additionally contain:

assay_name

raw_fluorescence

threshold

positive_detection

plate_run_id

If you select dataset = "unprocessed", the tibble will additionally contain:

assay_name

sample_type_name

raw_fluorescence

background_value

time

well_location

plate_run_id

Examples