Get Sample Details

process_well_sample_details reads in the plate run layout so sample identifiers can be mapped to the output of the Synergy H1 reader.

Usage

process_well_sample_details(
  filepath,
  sample_type,
  layout_type,
  single_assay_type,
  plate_run_id
)

Arguments

filepath

Synergy H1 reader output excel file with corresponding plate run layout in a sheet titled "Plate Map"

sample_type

either mucus or fin_clip

layout_type

either split_plate_early_late, split_plate_spring_winter, triplicate, or single_assay. If single_assay, you must specify single_assay_type.

plate_run_id

plate run identifier generated by running add_plate_run().

`single_assay_type`

assay type if running a plate with only one assay. Either OTS28 Early, OTS28 Late, OTS16 Spring, or OTS16 Winter. If not running a single assay for the plate, leave this argument empty.

Value

a table to be passed to process_sherlock():

plate_layout

location

sample_id

sample_type_id

assay_id

plate_run_id

Details

the output of this function will be used as the sample_details argument in process_sherlock() i.e. it will be used to connect the generic sample identifiers produced by the Synergy H1 reader to sample identifiers in the database. If the layout_type argument is split_plate, it divide the plate into columns 1:12 and 13:24 and assigns assays based on the following logic:

split_plate_early_late: columns 1:12 are OTS 28 Early and columns 13:24 are OTS 28 Late

split_plate_spring_winter: columns 1:12 are OTS 16 Spring and columns 13:24 are OTS 16 Winter

If the layout_type argument is triplicate, the assay ids are assigned as follows:

rows A, E, I, and M are OTS 28 Early

rows B, F, J, and N are OTS 28 Late

rows C, G, K, and O are OTS 16 Spring

rows D, H, L, and P are OTS 16 Winter

If the layout_type is single_assay, all sample IDs will be assigned the assay type you pass in as single_assay_type.