Set Status Code

Set the status code for existing samples

set_sample_status(con, sample_ids, status_code_id, comment = NULL)

Arguments

con

connection to the database

sample_ids

vector of sample ids to update

status

status code to use for the update

created status assigned automatically when a new sample ID is generated

prepped sample tube has been created

out to field sample tube is out in the field to be collected

return from field sample tube has returned from the field

in analysis sample is being analyzed

stored sample is in storage

archived sample has been sent to a tissue archive

other lab sample is out at another lab

See also

Examples

# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
#> Error: could not connect to server: Connection refused (0x0000274D/10061)
#> 	Is the server running on host "localhost" (::1) and accepting
#> 	TCP/IP connections on port 5432?
#> could not connect to server: Connection refused (0x0000274D/10061)
#> 	Is the server running on host "localhost" (127.0.0.1) and accepting
#> 	TCP/IP connections on port 5432?

sample_locations <- set_sample_locations(con,
                                         sample_ids = c("FTH_RM1722_3_A_1",
                                                         "FTH_RM1722_3_A_2"),
                                         status = "prepped",
                                         comment = "ready for pickup")
#> Error in set_sample_locations(con, sample_ids = c("FTH_RM1722_3_A_1",     "FTH_RM1722_3_A_2"), status = "prepped", comment = "ready for pickup"): could not find function "set_sample_locations"