Update Sample Location Status

update_sample_location_status() changes active flag on existing sample location in the sample location lookup table

update_sample_location_status(con, sample_location_id, set_active = TRUE)

Arguments

con

A DBI connection object obtained from DBI::dbConnect()

sample_location_id

A numeric ID for the targeted sample location get_sample_locations

set_active

A boolean, TRUE for activating and FALSE for deactivating. When a record is active, it is returned by default when get_sample_locations is called. This helps preserve look up values that are valid in historic contexts, but are no longer valid for current data records.

See also

Examples

# example database connection
cfg <- config::get()
con <- DBI::dbConnect(RPostgres::Postgres(),
                      dbname = cfg$dbname,
                      host = cfg$host,
                      port = cfg$port,
                      user = cfg$username,
                      password = cfg$password)
#> Error: could not connect to server: Connection refused (0x0000274D/10061)
#> 	Is the server running on host "localhost" (::1) and accepting
#> 	TCP/IP connections on port 5432?
#> could not connect to server: Connection refused (0x0000274D/10061)
#> 	Is the server running on host "localhost" (127.0.0.1) and accepting
#> 	TCP/IP connections on port 5432?

all_sample_locations <- get_sample_locations(con)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'dbObj' in selecting a method for function 'dbIsValid': object 'con' not found
View(all_sample_locations) # to view the ID of the sample location needing status change
#> Error in as.data.frame(x): object 'all_sample_locations' not found

#deactivate
update_sample_location_status(con, 4, set_active=FALSE)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'dbObj' in selecting a method for function 'dbIsValid': object 'con' not found
#reactivate
update_sample_location_status(con, 4)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'dbObj' in selecting a method for function 'dbIsValid': object 'con' not found